api_gateway/requirements.txt
Flask==2.0.2
grpcio-tools==1.27.1